// src/app/main-pages/financials/choose-earnings-chart/earnings-chart-option.enum.ts
/** A chart option can consist of one or more metrics */
export enum EarningsChartOption {
    REVENUE_VS_NET_INCOME = 'REVENUE_VS_NET_INCOME',
    REVENUE_VS_COST = 'REVENUE_VS_COST',
    REVENUE_VS_GROSS_PROFIT = 'REVENUE_VS_GROSS_PROFIT',

    INTEREST_INCOME = 'INTEREST_INCOME',
    STOCKHOLDERS_EQUITY = 'STOCKHOLDERS_EQUITY',

    OPERATING_INCOME = 'OPERATING_INCOME',
    GROSS_PROFIT_VS_SGA = 'GROSS_PROFIT_VS_SGA',
    OPERATIONS_VS_SGA = 'OPERATIONS_VS_SGA',
}